refactor: collapse duplicated bridge modules into single sources of truth
- Remove providerConfigBridge.ts (was a full copy of providerConfig.ts)
- Make providerConfig.ts use stateBridge directly, eliminating duplication
- Strip configBridge.ts to just setConfigProvider + readConfig (remove 4
  convenience wrappers that duplicated constants.ts functions)
- Inline readConfig() calls in RetryState.ts instead of wrapper functions
- Fix missed getConfig→readConfig rename in modelHandlerOpenAI.ts
- Re-export getToolUseMemoryEnabled from providerConfig in constants.ts

Net: -105 lines, 1 file deleted, no logic duplication.
